Reveal broken test
Chai’s `throw()` is a function, and must be called.  So neither
of these tests were doing anything.  The upgrade to mocha 5 caused the
last test to begin failing due to the uncaught error being thrown due
to the file not being found (reading the file is outside the await).